Rename Python mps_parser package to io
The cuopt.linear_programming.mps_parser package now exposes both
ParseMps and ParseLp, so 'mps_parser' is no longer accurate. Rename
the directory to 'io' (mirroring the cpp/include/cuopt/linear_programming/io/
layout on the C++ side) and rename the decorator
catch_mps_parser_exception → catch_io_exception.

This package was added on this branch in 72ba0540 (post-26.04) and has
not appeared in any release, so the rename is safe without a
compatibility shim.

Importing files that previously did 'from cuopt.linear_programming
import mps_parser' now import 'io as mps_parser' to keep the local
binding stable; user-facing docs/examples and the public top-level
import in cuopt/linear_programming/__init__.py use the new module
name directly. The C++ error tag MPS_PARSER_ERROR_TYPE that the
decorator parses out of RuntimeError messages is intentionally
unchanged — that tag is produced by the mps_parser_expects() macro,
which is shared between both parsers on the C++ side and would have
a much larger blast radius to rename.
