+++ docs/cuopt/source/hidden/mps-api.rst
@@ -5,9 +5,9 @@ cuOpt MPS/LP Parser API Reference
 MPS Parser
 ----------
 
-.. autofunction:: cuopt.linear_programming.mps_parser.ParseMps
+.. autofunction:: cuopt.linear_programming.io.ParseMps
 
 LP Parser
 ---------
 
-.. autofunction:: cuopt.linear_programming.mps_parser.ParseLp
+.. autofunction:: cuopt.linear_programming.io.ParseLp
