+++ docs/cuopt/source/hidden/mps-example.rst
@@ -9,5 +9,5 @@ Example
 .. code-block:: python
     :linenos:
 
-    from cuopt.linear_programming import mps_parser
-    x = mps_parser.ParseMps('good-mps-1.mps')
+    from cuopt.linear_programming import ParseMps
+    x = ParseMps('good-mps-1.mps')
