+++ python/cuopt/cuopt/CMakeLists.txt
@@ -5,7 +5,7 @@
 
 add_subdirectory(distance_engine)
 add_subdirectory(linear_programming/data_model)
-add_subdirectory(linear_programming/mps_parser)
+add_subdirectory(linear_programming/io)
 add_subdirectory(linear_programming/solver)
 add_subdirectory(routing)
