+++ python/cuopt/cuopt/linear_programming/__init__.py
@@ -3,7 +3,7 @@
 
 from cuopt.linear_programming import internals
 from cuopt.linear_programming.data_model import DataModel
-from cuopt.linear_programming.mps_parser import ParseLp, ParseMps
+from cuopt.linear_programming.io import ParseLp, ParseMps
 from cuopt.linear_programming.problem import Problem
 from cuopt.linear_programming.solution import Solution
 from cuopt.linear_programming.solver import BatchSolve, Solve
