+++ python/cuopt/cuopt/linear_programming/io/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from cuopt.linear_programming.mps_parser.parser import ParseLp, ParseMps, toDict
+from cuopt.linear_programming.io.parser import ParseLp, ParseMps, toDict
