+++ python/cuopt/cuopt/linear_programming/io/utilities/__init__.py
@@ -1,9 +1,9 @@
 # SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from cuopt.linear_programming.mps_parser.utilities.exception_handler import (
+from cuopt.linear_programming.io.utilities.exception_handler import (
     InputRuntimeError,
     InputValidationError,
     OutOfMemoryError,
-    catch_mps_parser_exception,
+    catch_io_exception,
 )
