+++ python/cuopt/cuopt/linear_programming/io/utilities/exception_handler.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2024-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
 import functools
@@ -17,7 +17,14 @@ class OutOfMemoryError(Exception):
     pass
 
 
-def catch_mps_parser_exception(f):
+def catch_io_exception(f):
+    """Translate the C++ parser's JSON-tagged RuntimeError to a typed Python
+    exception. The error tag string ("MPS_PARSER_ERROR_TYPE") is preserved
+    verbatim because it's produced by the C++ mps_parser_expects() macro,
+    which is shared between the MPS and LP parsers; renaming it would be a
+    C++-side change with a larger blast radius.
+    """
+
     @functools.wraps(f)
     def func(*args, **kwargs):
         try:
