+++ python/cuopt/cuopt/tests/linear_programming/test_cpu_only_execution.py
@@ -23,7 +23,7 @@ import subprocess
 import sys
 import time
 
-from cuopt.linear_programming import mps_parser
+from cuopt.linear_programming import io as mps_parser
 import pytest
 from cuopt import linear_programming
 from cuopt.linear_programming.solver.solver_parameters import CUOPT_TIME_LIMIT
@@ -301,7 +301,7 @@ def _run_in_subprocess(func, env=None, timeout=120):
 def _impl_lp_solve_cpu_only():
     """LP solve returns correctly-sized solution vectors."""
     from cuopt import linear_programming
-    from cuopt.linear_programming import mps_parser
+    from cuopt.linear_programming import io as mps_parser
 
     dataset_root = os.environ.get("RAPIDS_DATASET_ROOT_DIR", "./")
     mps_file = f"{dataset_root}/linear_programming/afiro_original.mps"
@@ -331,7 +331,7 @@ def _impl_lp_solve_cpu_only():
 def _impl_lp_dual_solution_cpu_only():
     """Dual solution and reduced costs are correctly sized."""
     from cuopt import linear_programming
-    from cuopt.linear_programming import mps_parser
+    from cuopt.linear_programming import io as mps_parser
 
     dataset_root = os.environ.get("RAPIDS_DATASET_ROOT_DIR", "./")
     mps_file = f"{dataset_root}/linear_programming/afiro_original.mps"
@@ -364,7 +364,7 @@ def _impl_mip_solve_cpu_only():
     from cuopt.linear_programming.solver.solver_parameters import (
         CUOPT_TIME_LIMIT,
     )
-    from cuopt.linear_programming import mps_parser
+    from cuopt.linear_programming import io as mps_parser
 
     dataset_root = os.environ.get("RAPIDS_DATASET_ROOT_DIR", "./")
     mps_file = f"{dataset_root}/mip/bb_optimality.mps"
@@ -400,7 +400,7 @@ def _impl_warmstart_cpu_only():
         CUOPT_PRESOLVE,
     )
     from cuopt.linear_programming.solver_settings import SolverMethod
-    from cuopt.linear_programming import mps_parser
+    from cuopt.linear_programming import io as mps_parser
 
     dataset_root = os.environ.get("RAPIDS_DATASET_ROOT_DIR", "./")
     mps_file = f"{dataset_root}/linear_programming/afiro_original.mps"
