+++ python/cuopt/cuopt/tests/linear_programming/test_lp_solver.py
@@ -3,7 +3,7 @@
 
 import os
 
-from cuopt.linear_programming import mps_parser
+from cuopt.linear_programming import io as mps_parser
 import numpy as np
 import pytest
